+++ docnexus/__init__.py
@@ -2,7 +2,7 @@
 DocNexus - The Ultimate All-in-One Document Engine
 """
 
-__version__ = '1.1.3'
-__author__ = 'OmniDoc Maintainers'
-__email__ = 'maintainers@omnidoc.org'
-__description__ = 'OmniDoc - The Ultimate All-in-One Document Engine'
+__version__ = '1.1.4'
+__author__ = 'Doc Nexus Maintainers'
+__email__ = 'maintainers@docnexus.org'
+__description__ = 'DocNexus - The Ultimate All-in-One Document Engine'